# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Kenya Location Intelligence API",
    "description": "Geospatial API for Kenya administrative boundaries, places, nearby searches, and reverse geolocation.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/locations/search": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Search Kenya administrative locations",
        "description": "Searches counties, sub-counties, and wards by name or administrative level.",
        "operationId": "SearchLocations",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Optional text used to search by name.",
            "schema": {
              "type": "string",
              "description": "Optional text used to search by name."
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Administrative level: 1 = county, 2 = sub-county, 3 = ward.",
            "schema": {
              "type": "integer",
              "description": "Administrative level: 1 = county, 2 = sub-county, 3 = ward.",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
            "schema": {
              "type": "integer",
              "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationSearchResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locations/{id}": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get an administrative location",
        "description": "Returns details for a county, sub-county, or ward by its identifier.",
        "operationId": "GetLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Administrative location identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Administrative location identifier."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locations/{id}/hierarchy": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get the administrative hierarchy",
        "description": "Returns the county-to-ward hierarchy containing the requested administrative location.",
        "operationId": "GetLocationHierarchy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Administrative location identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Administrative location identifier."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationHierarchyItemResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nearby": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Find nearby administrative locations",
        "description": "Finds administrative locations near a latitude and longitude using a radius in kilometres.",
        "operationId": "GetNearbyLocations",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "description": "Latitude in decimal degrees. Valid range: -90 to 90.",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude in decimal degrees. Valid range: -90 to 90.",
              "format": "double"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "description": "Longitude in decimal degrees. Valid range: -180 to 180.",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude in decimal degrees. Valid range: -180 to 180.",
              "format": "double"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "description": "Search radius in kilometres. Default: 25. Valid range: 0.1 to 250.",
            "schema": {
              "type": "number",
              "description": "Search radius in kilometres. Default: 25. Valid range: 0.1 to 250.",
              "format": "double"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
            "schema": {
              "type": "integer",
              "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NearbyLocationResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApiNotFoundError": {
        "required": [
          "error",
          "id"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "LocationDetailResponse": {
        "required": [
          "code",
          "id",
          "latitude",
          "level",
          "levelName",
          "longitude",
          "name",
          "source",
          "sourceVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "levelName": {
            "type": "string"
          },
          "parent": {
            "$ref": "#/components/schemas/LocationParentResponse"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "type": "string"
          },
          "sourceVersion": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LocationHierarchyItemResponse": {
        "required": [
          "code",
          "depth",
          "id",
          "level",
          "levelName",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "levelName": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LocationParentResponse": {
        "required": [
          "code",
          "id",
          "level",
          "levelName",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "levelName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LocationSearchResponse": {
        "required": [
          "code",
          "id",
          "latitude",
          "level",
          "levelName",
          "longitude",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "levelName": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NearbyLocationResponse": {
        "required": [
          "code",
          "distanceMeters",
          "id",
          "latitude",
          "level",
          "levelName",
          "longitude",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "levelName": {
            "type": "string"
          },
          "distanceMeters": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}